add img Signed-off-by: lingfe <13068326391@163.com>

+++ zber_sys/src/com/yyf/controller/ImagesController.java
@@ -1,7 +1,9 @@
 package com.yyf.controller;
 
+import java.io.File;
 import java.io.IOException;
 import java.util.Date;
+import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
@@ -16,6 +18,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.yyf.controller.util.SYS_GET;
 import com.yyf.controller.util.UploadUtils;
 import com.yyf.model.JosnModel;
 import com.yyf.model.Tab_images;
@@ -30,6 +33,46 @@ public class ImagesController {
 	
 	/**
 	 * 
+	 * 根据setId获取图片数据
+	 * @author lingfe     
+	 * @created 2019年3月25日 下午3:47:07  
+	 * @param setId
+	 * @return
+	 */
+	@RequestMapping(value = "/getWhereSetId", method = {RequestMethod.POST,RequestMethod.GET})
+    @ResponseBody
+	public JosnModel<Object> getWhereSetId(@RequestParam(value="setId",required=false)String setId){
+		//实例化对象
+		JosnModel<Object> josn=new JosnModel<Object>();
+		if(!StringUtils.isEmpty(setId)){
+			//执行删除
+			List<Tab_images> list=iimagesService.getWhereLbtAttributeId(setId);
+			if(list!=null && list.size()>0){
+				for (int j = 0; j < list.size(); j++) {
+					Tab_images tab_images=list.get(j);
+					//验证图片路径
+					if(tab_images.getImgUrl().indexOf("http") ==-1){
+    					String img= SYS_GET.GET_IMG_PATH_URL + tab_images.getImgUrl();
+    					tab_images.setImgUrl(img);
+    				}
+					
+					//数据更新
+					list.set(j, tab_images);
+				}
+			}
+			
+			josn.state=200;
+			josn.msg="查询成功!";
+			josn.data=list;
+		}else{
+			josn.msg="查询失败!id为空!";
+		}
+		
+		return josn;
+	}
+	
+	/**
+	 * 
 	 * 根据图片id删除图片
 	 * @author lingfe     
 	 * @created 2018年10月31日 下午7:17:42  
@@ -38,17 +81,33 @@ public class ImagesController {
 	 */
 	@RequestMapping(value = "/deleteWhereId", method = {RequestMethod.POST,RequestMethod.GET})
     @ResponseBody
-    public JosnModel<Object> del(@RequestParam(value="id",required=false)String id){
+    public JosnModel<Object> deleteWhereId(@RequestParam(value="id",required=false)String id){
 		//实例化对象
 		JosnModel<Object> josn=new JosnModel<Object>();
 		if(!StringUtils.isEmpty(id)){
-			//执行删除
-			int tt=iimagesService.deleteWhereId(id);
-			if(tt>=1){
-				josn.state=200;
-				josn.msg="删除成功!";
+			//执行查询是否存在
+			Tab_images img=iimagesService.getWhereId(id);
+			if(img!=null){
+				//执行删除
+				int tt=iimagesService.deleteWhereId(id);
+				if(tt>=1){
+					
+					//验证图片路径
+					if(img.getFull_path()!=null&&!"".equals(img.getFull_path())){
+						//删除文件,带路径的文件名
+				        File file = new File(img.getFull_path());
+				        if(file.delete()){
+				            System.out.println("deleted path="+img.getFull_path());
+				        }
+					}
+					
+			        josn.state=200;
+					josn.msg="删除成功!";
+				}else{
+					josn.msg="删除失败!";	
+				}
 			}else{
-				josn.msg="删除失败!";	
+				josn.msg="图片信息不存在!或者id无效!";
 			}
 		}else{
 			josn.msg="删除失败!id为空!";
